% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/withDoc6.R
=======
% Please edit documentation in R/withDoc.R
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
\docType{data}
\name{cbc.rrbs}
\alias{cbc.rrbs}
\title{Known cell type mixture for tissue (bisulfite sequencing)}
\format{An object of class \code{data.frame} with 35 rows and 3 columns.}
\usage{
cbc.rrbs
}
\description{
Known cell type mixture for tissue (bisulfite sequencing)
}
\keyword{datasets}
