% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{celltypes.rrbs}
\alias{celltypes.rrbs}
\title{vector of kinds of cell types (bisulfite sequencing)}
\format{
An object of class \code{character} of length 4.
}
\usage{
celltypes.rrbs
}
\description{
vector of kinds of cell types (bisulfite sequencing)
}
\keyword{datasets}
