% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/withDoc6.R
=======
% Please edit documentation in R/withDoc.R
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
\docType{data}
\name{cnts.celltypes}
\alias{cnts.celltypes}
\title{Count matrix for cell types (RNAseq)}
\format{An object of class \code{matrix} with 500 rows and 64 columns.}
\usage{
cnts.celltypes
}
\description{
Count matrix for cell types (RNAseq)
}
\keyword{datasets}
