% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{cnts.sc.G1}
\alias{cnts.sc.G1}
\title{count matrix for filtered scRNAseq dataset for cell phase "G1"}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3439 rows and 638 columns.
}
\usage{
cnts.sc.G1
}
\description{
count matrix for filtered scRNAseq dataset for cell phase "G1"
}
\keyword{datasets}
