% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{cnts.sc}
\alias{cnts.sc}
\title{count matrix for filtered scRNAseq dataset}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3439 rows and 1747 columns.
}
\usage{
cnts.sc
}
\description{
count matrix for filtered scRNAseq dataset
}
\keyword{datasets}
