% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{cnts.scrnaseq}
\alias{cnts.scrnaseq}
\title{Count matrix for single cell RNAseq (scRNAseq)}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3439 rows and 1749 columns.
}
\usage{
cnts.scrnaseq
}
\description{
Count matrix for single cell RNAseq (scRNAseq)
}
\keyword{datasets}
