% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc (Autosaved).R, R/withDoc.R
\docType{data}
\name{cnts.tissue}
\alias{cnts.tissue}
\title{Count matrix for tissue (RNAseq)}
\format{
An object of class \code{matrix} with 500 rows and 35 columns.

An object of class \code{matrix} with 500 rows and 35 columns.
}
\usage{
cnts.tissue

cnts.tissue
}
\description{
Count matrix for tissue (RNAseq)

Count matrix for tissue (RNAseq)
}
\keyword{datasets}
