% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/withDoc6.R
=======
% Please edit documentation in R/withDoc.R
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
\docType{data}
\name{data_celltypes_rnaseq}
\alias{data_celltypes_rnaseq}
\title{This is the example cell type data for RNA sequencing}
\usage{
data("data_celltypes_rnaseq")
}
\description{
\itemize{
<<<<<<< HEAD
\item dge.celltypes DGE object from (EdgeR) of RNAseq results
=======
\item dge.celltypes DGE object (from EdgeR) of RNAseq results
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
\item cnts.celltypes count matrix
\item design.rnaseq design matrix
\item sample.id.rnaseq sample IDs
}
}
\keyword{datasets}
