% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{data_celltypes_rrbs}
\alias{data_celltypes_rrbs}
\title{This is the example cell type data for bisulfite sequencing}
\usage{
data("data_celltypes_rrbs")
}
\description{
\itemize{
\item methmat methylation matrix
\item celltypes.rrbs vector of kinds of cell types
\item design.rrbs design matrix
\item sample.id.rrbs sample IDs
}
}
\keyword{datasets}
