% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{data_scrnaseq}
\alias{data_scrnaseq}
\title{This is the example data for scRNA sequencing}
\usage{
data("data_scrnaseq")
}
\description{
\itemize{
\item cnts.scrnaseq count matrix for full scRNAseq dataset
}
}
\keyword{datasets}
