% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{data_tissue_rrbs}
\alias{data_tissue_rrbs}
\title{This is the example tissue data for bisulfite sequencing}
\usage{
data("data_tissue_rrbs")
}
\description{
\itemize{
\item methmat.tissue methylation matrix
\item sample.id.tissue sample IDs
\item cbc.rrbs known cell type mixture for tissue
}
}
\keyword{datasets}
