% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/withDoc6.R
=======
% Please edit documentation in R/withDoc.R
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
\docType{data}
\name{design.rnaseq}
\alias{design.rnaseq}
\title{Design matrix for cell types (RNAseq)}
\format{An object of class \code{matrix} with 64 rows and 4 columns.}
\usage{
design.rnaseq
}
\description{
Design matrix for cell types (RNAseq)
}
\keyword{datasets}
