% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/withDoc6.R
=======
% Please edit documentation in R/withDoc.R
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
\docType{data}
\name{design.rrbs}
\alias{design.rrbs}
\title{Design matrix for cell types (bisulfite sequencing)}
\format{An object of class \code{data.frame} with 24 rows and 4 columns.}
\usage{
design.rrbs
}
\description{
Design matrix for cell types (bisulfite sequencing)
}
\keyword{datasets}
