% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{design_s.sc}
\alias{design_s.sc}
\title{Design matrix (scRNAseq)}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 238 rows and 2 columns.
}
\usage{
design_s.sc
}
\description{
Design matrix (scRNAseq)
}
\keyword{datasets}
