% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{dge.sc}
\alias{dge.sc}
\title{DGE for training set (scRNAseq)}
\format{
An object of class \code{DGEList} with 3439 rows and 400 columns.
}
\usage{
dge.sc
}
\description{
DGE for training set (scRNAseq)
}
\keyword{datasets}
