% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc (Autosaved).R, R/withDoc.R
\docType{data}
\name{dge.tissue}
\alias{dge.tissue}
\title{DGE object (from EdgeR) of RNAseq results for tissue}
\format{
An object of class \code{DGEList} with 500 rows and 35 columns.

An object of class \code{DGEList} with 500 rows and 35 columns.
}
\usage{
dge.tissue

dge.tissue
}
\description{
DGE object (from EdgeR) of RNAseq results for tissue

DGE object (from EdgeR) of RNAseq results for tissue
}
\keyword{datasets}
