% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/withDoc6.R
=======
% Please edit documentation in R/withDoc.R
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
\name{getb0.biseq}
\alias{getb0.biseq}
\title{compute b0, projection matrix, given methylation counts}
\usage{
<<<<<<< HEAD
getb0.biseq(methmat, design, sigg = NULL)
}
\arguments{
\item{methmat}{a matrix of counts with rows corresponding to methylation sites and columns. Columns are: chr start   end strand coverage1 numCs1 numTs1 coverage2 numCs2 numTs2 ....}

\item{design}{design matrix, output of model.matrix}

\item{sigg}{predetermined signature CpG sites. Format sites as chromosome
name, chromosome location, strand: chrN_position(+/-). For example, chr1_906825-}
=======
getb0.biseq(methmat, design)
}
\arguments{
\item{methmat}{a matrix of counts with rows corresponding to methylation
sites and columns. Columns are: chr start   end strand coverage1 numCs1
numTs1 coverage2 numCs2 numTs2 ....}

\item{design}{design matrix, output of model.matrix}
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
}
\value{
b0 projection matrix. Coefficients are beta.
}
\description{
compute b0, projection matrix, given methylation counts
}
