% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc2.R
\name{getb0}
\alias{getb0}
\title{compute B0, projection matrix, given counts}
\usage{
getb0(countmat, design, ncpm.min = 1, nsamp.min = 4)
}
\arguments{
\item{countmat}{a matrix of counts with rows corresponding to genes and columns to samples}

\item{design}{output of model.matrix}

\item{ncpm.min}{filtering threshold for cpm}

\item{nsamp.min}{minimum number of samples that must have cpm exceeding ncpm.min for a gene to be retained}
}
\value{
b0 projection matrix

pfstat f-stat p-values for each gene

dge DGE object
}
\description{
compute B0, projection matrix, given counts
}
\examples{
data("leucegene_btmg_b0_counts")
okids = sample.idVec()
cnts = leucegene_btmg_b0_counts[, okids]
tissf = factor(tissueVec(), levels=celltypeVec())
des = model.matrix(~-1+tissf)
b0 = getb0(cnts, des, ncpm.min=1, nsamp.min=4)
summary(b0)
}
