% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc6.R
\name{getcellcycle}
\alias{getcellcycle}
\title{filter for cells with particular cell cycle phase: "G1", "G2M", or "S"}
\usage{
getcellcycle(scrna_mat, cellcycle = "G1", species = "human")
}
\arguments{
\item{scrna_mat}{count matrix  for single cell RNAseq with ensembl gene IDs}

\item{cellcycle}{filter for specific cell cycle phase: "G1", "G2M", or "S". Default is NULL, for no cell cycle phase filtering.}

\item{species}{"human" or "mouse"}
}
\value{
count matrix containing only cells with the selected cell cycle phase
}
\description{
filter for cells with particular cell cycle phase: "G1", "G2M", or "S"
}
