% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconvSeq2.R, R/withDoc.R
\name{getcorr}
\alias{getcorr}
\alias{getcorr}
\title{compute correlation between estimated and actual composition}
\usage{
getcorr(x1, x2)

getcorr(x1, x2)
}
\arguments{
\item{x1}{composition matrix where rows are samples and columns are cell types}

\item{x2}{second composition matrix where rows and columns match x1}

\item{x1}{composition matrix where rows are samples and columns are cell types}

\item{x2}{second composition matrix where rows and columns match x1}
}
\value{
correlation by sample

cor.type correlation by cell type

cor.sample correlation by sample
}
\description{
compute correlation between estimated and actual composition

compute correlation between estimated and actual composition
}
\examples{
#RNA sequencing example

set.seed(1234)
data("data_celltypes_rnaseq") #contains DGE object dge.celltypes, count matrix cnts.celltypes, and design matrix des for cell types design.rnaseq, sample.id.rnaseq
dge.celltypes = getdge(cnts.celltypes, design.rnaseq, ncpm.min=1, nsamp.min=4)
b0 = getb0.rnaseq(dge.celltypes, design.rnaseq, ncpm.min=1, nsamp.min=4)
nb0 = 50
resultx1 = getx1.rnaseq(nb0,b0,dge.celltypes)
x2 = as.data.frame(design.rnaseq,row.names=sample.id.rnaseq)
getcorr(resultx1$x1,x2)

#Validate projection matrix on whole tissue
data("data_tissue_rnaseq") #contains DGE object dge.tissue, count matrix cnts.tissue, and CBC cbc.rnaseq
nb0=50
dge.tissue = getdge(cnts.tissue,design=NULL, ncpm.min=1, nsamp.min=4)
resultx1.tissue = getx1.rnaseq(nb0,b0, dge.tissue)
x1 = cbind(lymph=resultx1.tissue$x1$tissuefBcell+resultx1.tissue$x1$tissuefTcell, mono = resultx1.tissue$x1$tissuefMonocytes, gran = resultx1.tissue$x1$tissuefGranulocytes)
x2 = as.matrix(cbc.rnaseq/100)
getcorr(x1,x2)

#Bisulfite sequencing example

set.seed(1234)
data("data_celltypes_rrbs") #contains methylation matrix methmat, design.rrbs, celltypes.rrbs, sample.id.rrbs
b0 = getb0.biseq(methmat, design.rrbs)
nb0=250
resultx1 = getx1.biseq(nb0,b0,methmat,sample.id.rrbs,celltypes.rrbs)
x2 = as.data.frame(design.rrbs,row.names=sample.id.rrbs)
getcorr(resultx1$x1,x2)

#Validate projection matrix on whole tissue
data("data_tissue_rrbs") #contains methylation matrix methmat.tissue, sample.id.tissue, actual differential cell count cbc.rrbs
nb0=250
resultx1.tissue = getx1.biseq(nb0,b0,methmat.tissue,sample.id.tissue,celltypes.rrbs)
x1 = cbind(lymph=resultx1.tissue$x1[,1]+resultx1.tissue$x1[,2], mono = resultx1.tissue$x1[,3], gran = resultx1.tissue$x1[,4])
x2 = as.matrix(cbc.rrbs/100)
getcorr(x1,x2)
#RNA sequencing example

set.seed(1234)
data("celltypes_cnts")
okids = sample.idVec()
cnts = celltypes_cnts[, okids]
tissf = factor(tissueVec(), levels=celltypeVec())
des = model.matrix(~-1+tissf)
b0 = getb0.rnaseq(cnts, des, ncpm.min=1, nsamp.min=4)
nb0 = 50
dge = getdge(cnts, des, ncpm.min=1, nsamp.min=4)
resultx1 = getx1.rnaseq(nb0,b0,dge)
x2 = as.data.frame(des,row.names=okids)
getcorr(resultx1$x1,x2)

#Validate projection matrix on whole tissue
data("wholeblood_cnts")
okids.tissue = sample.id.tissue.Vec()
cnts.tissue = wholeblood_cnts[,okids.tissue]
nb0=3000
dge.tissue = getdge(cnts.tissue,design=NULL, ncpm.min=1, nsamp.min=4)
resultx1 = getx1.rnaseq(nb0,b0, dge.tissue)
cbc = cbcDF()
x1 = cbind(lymph=resultx1$x1[as.character(cbc$sample.id),1]+resultx1$x1[as.character(cbc$sample.id),2], mono = resultx1$x1[as.character(cbc$sample.id),3], gran = resultx1$x1[as.character(cbc$sample.id),4])
x2=as.matrix(cbc[,2:4])/100
getcorr(x1,x2)

#Bisulfite sequencing example

set.seed(1234)
data("methmat")
data("methmat_geo_small")
methmat = methmat_geo
okids = sample.idVec2()
tissf = factor(tissueVec2(), levels=celltypeVec2())
des = model.matrix(~-1+tissf)
b0 = getb0.biseq(methmat, des)
nb0=1000
resultx1 = getx1.biseq(nb0,b0,methmat,okids,celltypeVec2())
x2 = as.data.frame(des,row.names=okids)
getcorr(resultx1$x1,x2)

#Validate projection matrix on whole tissue
data("methmat_wholeblood_small")
methmat.tissue = methmat_wholeblood
okids.tissue = sample.id.tissue.Vec2()
nb0=1000
resultx1 = getx1.biseq(nb0,b0,methmat.tissue,okids.tissue,celltypeVec2())
cbc = cbcDF()
x1 = cbind(lymph=resultx1$x1[as.character(cbc$sample.id),1]+resultx1$x1[as.character(cbc$sample.id),2], mono = resultx1$x1[as.character(cbc$sample.id),3], gran = resultx1$x1[as.character(cbc$sample.id),4])
x2=as.matrix(cbc[,2:4])/100
getcorr(x1,x2)
}
