% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/withDoc6.R
=======
% Please edit documentation in R/withDoc.R
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
\name{getdge}
\alias{getdge}
\title{obtain EdgeR DGE object from count matrix}
\usage{
<<<<<<< HEAD
getdge(countmat, design, ncpm.min = 1, nsamp.min = 4,
  method = "auto")
=======
getdge(countmat, design, ncpm.min = 1, nsamp.min = 4)
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
}
\arguments{
\item{countmat}{a matrix of counts with rows corresponding to
genes and columns to samples}

\item{design}{output of model.matrix}

\item{ncpm.min}{filtering threshold for cpm}

<<<<<<< HEAD
\item{nsamp.min}{minimum number of samples that must have cpm}

\item{method}{method used to esimate trended dispersion in EdgeR: "auto" (default), "bin.spline", "bin.loess", "power", "spline"
=======
\item{nsamp.min}{minimum number of samples that must have cpm
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
exceeding ncpm.min for a gene to be retained}
}
\value{
dge DGE object
}
\description{
obtain EdgeR DGE object from count matrix
}
