% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\name{getmethmat}
\alias{getmethmat}
\title{makes methylation count matrix using output from BSMAP. Removes sites with 0 counts in all samples.
Input file (from BSMAP) has columns: chromosome, position, strand, context, ratio, eff_CT_count, C_count, CT_count, rev_G_count, rev_GA_count, CI_lower, CI_upper
Sample input files: sample1_methratio.txt, sample1_mehtratio.txt
Extract CpGs from input files and output in filname.CpG
Add "Chr" to chromosome name and output in filname.CpG_chr}
\usage{
getmethmat(filnames, sample.id)
}
\arguments{
\item{filnames}{input filenames}

\item{sample.id}{sample IDs}
}
\value{
methylation count matrix where rows are CpG sites and columns are: chromosome, start, end, strand, number of Ts+Cs for sample 1, number of Cs for sample 1, number of Ts for sample 1, ....
}
\description{
makes methylation count matrix using output from BSMAP. Removes sites with 0 counts in all samples.
Input file (from BSMAP) has columns: chromosome, position, strand, context, ratio, eff_CT_count, C_count, CT_count, rev_G_count, rev_GA_count, CI_lower, CI_upper
Sample input files: sample1_methratio.txt, sample1_mehtratio.txt
Extract CpGs from input files and output in filname.CpG
Add "Chr" to chromosome name and output in filname.CpG_chr
}
\examples{
 file1 = system.file("extdata","sample1_methratio.txt", package="deconvSeq")
 file2 = system.file("extdata","sample2_methratio.txt", package="deconvSeq")
methmat = getmethmat(filnames=c(file1,file2), sample.id=c("sample1","sample2"))
}
