% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc (Autosaved).R, R/withDoc.R
\name{getrnamat}
\alias{getrnamat}
\title{Combine count matrix for individual samples}
\usage{
getrnamat(filnames, sample.id)

getrnamat(filnames, sample.id)
}
\arguments{
\item{filnames}{filenames of individual samples}

\item{sample.id}{sample IDs}
}
\value{
count matrix with genes in rows. First column is gene name and

count matrix with genes in rows. First column is gene name and
}
\description{
Takes count matrix for individual samples (e.g. output of HTSeq output). Format: first column is Ensembl gene id, second column is count, no header. Removes genes with all 0 reads and rows that are not Ensembl gene IDs. Output is a count matrix where rows are genes and columns are samples.
Sample input files:  sample_genecounts1.txt, sample_genecounts2.txt

Takes count matrix for individual samples (e.g. output of HTSeq output). Format: first column is Ensembl gene id, second column is count, no header. Removes genes with all 0 reads and rows that are not Ensembl gene IDs. Output is a count matrix where rows are genes and columns are samples.
Sample input files:  sample_genecounts1.txt, sample_genecounts2.txt
}
\examples{
file1 = system.file("extdata","sample1_genecounts.txt", package="deconvSeq")
file2 = system.file("extdata","sample2_genecounts.txt", package="deconvSeq")
countmat = getrnamat(filnames=c(file1,file2),sample.id=c("sample1","sample2"))
file1 = system.file("extdata","sample1_genecounts.txt", package="deconvSeq")
file2 = system.file("extdata","sample2_genecounts.txt", package="deconvSeq")
countmat = getrnamat(filnames=c(file1,file2),sample.id=c("sample1","sample2"))
}
