% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc2.R
\name{getx1}
\alias{getx1}
\title{compute mixture data given projection matrix}
\usage{
getx1(NB0, resultb0, dge, MAXITER = 1000)
}
\arguments{
\item{NB0}{number of genes to be retained, ordered by f-stat p-value}

\item{resultb0}{output of \code{\link{getb0}}}

\item{dge}{output of \code{\link{getdge}}}

\item{MAXITER}{integer number of iterations allowed}
}
\value{
x1 cell mixture of sample

converged convergence
}
\description{
compute mixture data given projection matrix
}
