% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\name{getx1.rnaseq}
\alias{getx1.rnaseq}
\title{compute mixture data given projection matrix (RNAseq)}
\usage{
getx1.rnaseq(NB0, resultb0, dge_s, MAXITER = 1000, x0 = NULL)
}
\arguments{
\item{NB0}{number of genes to be retained, ordered by f-stat p-value}

\item{resultb0}{output of \code{\link{getb0.rnaseq}}}

\item{dge_s}{output of \code{\link{getdge}}}

\item{MAXITER}{integer number of iterations allowed}

\item{x0}{initial cell type composition for fitting}
}
\value{
x1 cell mixture of sample

converged convergence. 1=converged
}
\description{
compute mixture data given projection matrix (RNAseq)
}
