% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc (Autosaved).R, R/withDoc.R
\docType{data}
\name{methmat}
\alias{methmat}
\title{methylation matrix for cell types (bisulfite sequencing)}
\format{
An object of class \code{data.frame} with 500 rows and 76 columns.

An object of class \code{data.frame} with 500 rows and 76 columns.
}
\usage{
methmat

methmat
}
\description{
methylation matrix for cell types (bisulfite sequencing)

methylation matrix for cell types (bisulfite sequencing)
}
\keyword{datasets}
