% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_diffMethFromDesign_matrix.R
\name{my_diffMethFromDesign_matrix}
\alias{my_diffMethFromDesign_matrix}
\title{modified from methylKit, performs regression given methylation matrix}
\usage{
my_diffMethFromDesign_matrix(cur_methmat, cur_design, cur_formula)
}
\arguments{
\item{cur_methmat}{methylation matrix}

\item{cur_design}{design matrix}

\item{cur_formula}{formula}
}
\value{
result
}
\description{
modified from methylKit, performs regression given methylation matrix
}
