% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{resultx1}
\alias{resultx1}
\title{Results from celltype analysis (RNAseq)}
\format{
An object of class \code{list} of length 2.
}
\usage{
resultx1
}
\description{
Results from celltype analysis (RNAseq)
}
\keyword{datasets}
