% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{resultx1.tissue}
\alias{resultx1.tissue}
\title{Results from tissue analysis (RNAseq)}
\format{
An object of class \code{list} of length 2.
}
\usage{
resultx1.tissue
}
\description{
Results from tissue analysis (RNAseq)
}
\keyword{datasets}
