% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{resultx1.tissue.rrbs}
\alias{resultx1.tissue.rrbs}
\title{Results from tissue analysis (bisulfite sequencing)}
\format{
An object of class \code{list} of length 2.
}
\usage{
resultx1.tissue.rrbs
}
\description{
Results from tissue analysis (bisulfite sequencing)
}
\keyword{datasets}
