% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{resultx1_s.sc}
\alias{resultx1_s.sc}
\title{Results (scRNAseq)}
\format{
An object of class \code{list} of length 2.
}
\usage{
resultx1_s.sc
}
\description{
Results (scRNAseq)
}
\keyword{datasets}
