% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/withDoc6.R
=======
% Please edit documentation in R/withDoc.R
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
\docType{data}
\name{sample.id.rnaseq}
\alias{sample.id.rnaseq}
\title{Sample IDs for cell types (RNAseq)}
\format{An object of class \code{character} of length 64.}
\usage{
sample.id.rnaseq
}
\description{
Sample IDs for cell types (RNAseq)
}
\keyword{datasets}
