% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/withDoc6.R
=======
% Please edit documentation in R/withDoc.R
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
\docType{data}
\name{sample.id.tissue}
\alias{sample.id.tissue}
\title{Sample IDs for tissue (bisulfite sequencing)}
\format{An object of class \code{character} of length 35.}
\usage{
sample.id.tissue
}
\description{
Sample IDs for tissue (bisulfite sequencing)
}
\keyword{datasets}
