% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{sample.id.tissue.rrbs}
\alias{sample.id.tissue.rrbs}
\title{Sample IDs for tissue (bisulfite sequencing)}
\format{
An object of class \code{character} of length 35.
}
\usage{
sample.id.tissue.rrbs
}
\description{
Sample IDs for tissue (bisulfite sequencing)
}
\keyword{datasets}
