% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\name{sample.id.tissue.Vec2}
\alias{sample.id.tissue.Vec2}
\title{generate ordered sample identifiers for RNAseq data for whole blood}
\usage{
sample.id.tissue.Vec2()
}
\description{
generate ordered sample identifiers for RNAseq data for whole blood
}
