% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{sample.id_s.sc}
\alias{sample.id_s.sc}
\title{sample IDs (scRNAseq)}
\format{
An object of class \code{character} of length 238.
}
\usage{
sample.id_s.sc
}
\description{
sample IDs (scRNAseq)
}
\keyword{datasets}
