% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{cbc.rnaseq}
\alias{cbc.rnaseq}
\title{Known cell type mixture for tissue (RNAseq)}
\format{An object of class \code{data.frame} with 35 rows and 3 columns.}
\usage{
cbc.rnaseq
}
\description{
Known cell type mixture for tissue (RNAseq)
}
\keyword{datasets}
