% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc (Autosaved).R, R/withDoc.R
\docType{data}
\name{cnts.celltypes}
\alias{cnts.celltypes}
\title{Count matrix for cell types (RNAseq)}
\format{
An object of class \code{matrix} with 500 rows and 64 columns.

An object of class \code{matrix} with 500 rows and 64 columns.
}
\usage{
cnts.celltypes

cnts.celltypes
}
\description{
Count matrix for cell types (RNAseq)

Count matrix for cell types (RNAseq)
}
\keyword{datasets}
