% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc6.R
\docType{data}
\name{cnts.sc.G1}
\alias{cnts.sc.G1}
\title{Count matrix for single cell RNAseq after filtering cnts.sc for G1 cell cycle phase}
\format{An object of class \code{matrix} with 3439 rows and 496 columns.}
\usage{
cnts.sc.G1
}
\description{
Count matrix for single cell RNAseq after filtering cnts.sc for G1 cell cycle phase
}
\keyword{datasets}
