% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc (Autosaved).R, R/withDoc.R
\docType{data}
\name{cnts.sc}
\alias{cnts.sc}
\title{Count matrix for single cell RNAseq after filtering for count threshold of 0.05}
\format{
An object of class \code{matrix} with 3439 rows and 1747 columns.

An object of class \code{matrix} with 3439 rows and 1747 columns.
}
\usage{
cnts.sc

cnts.sc
}
\description{
Count matrix for single cell RNAseq after filtering for count threshold of 0.05

Count matrix for single cell RNAseq after filtering for count threshold of 0.05
}
\keyword{datasets}
