% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc6.R
\docType{data}
\name{cnts.scrnaseq}
\alias{cnts.scrnaseq}
\title{Count matrix for single cell RNAseq (scRNAseq)}
\format{An object of class \code{matrix} with 23398 rows and 1749 columns.}
\usage{
cnts.scrnaseq

cnts.scrnaseq
}
\description{
Count matrix for single cell RNAseq (scRNAseq)

Count matrix for single cell RNAseq
}
\keyword{datasets}
