% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/withDoc6.R
=======
% Please edit documentation in R/withDoc.R
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
\docType{data}
\name{cnts.tissue}
\alias{cnts.tissue}
\title{Count matrix for tissue (RNAseq)}
\format{An object of class \code{matrix} with 500 rows and 35 columns.}
\usage{
cnts.tissue
}
\description{
Count matrix for tissue (RNAseq)
}
\keyword{datasets}
