% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{data_celltypes_rnaseq}
\alias{data_celltypes_rnaseq}
\title{This is the example cell type data for RNA sequencing}
\usage{
data("data_celltypes_rnaseq")
}
\description{
\itemize{
\item dge.celltypes DGE object (from EdgeR) of RNAseq results
\item cnts.celltypes count matrix
\item design.rnaseq design matrix
\item sample.id.rnaseq sample IDs
}
}
\keyword{datasets}
