% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{data_scrnaseq}
\alias{data_scrnaseq}
\title{This is the example data for scRNA sequencing}
\usage{
data("data_scrnaseq")
}
\description{
\itemize{
\item cnts.scrnaseq count matrix for full scRNAseq dataset
\item cnts.sc count matrix for filtered scRNAseq dataset
\item cnts.sc.G1 count matrix for filtered scRNAseq dataset for cell phase "G1"
\item b0.sc projection matrix
\item dge.sc DGE for training set
}
}
\keyword{datasets}
