% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{data_tissue_rnaseq}
\alias{data_tissue_rnaseq}
\title{This is the example tissue data for RNA sequencing}
\usage{
data("data_tissue_rnaseq")
}
\description{
\itemize{
\item dge.tissue DGE object (from EdgeR) of RNAseq results
\item cnts.tissue count matrix
\item cbc.rnaseq known cell type mixture for tissue
\item resultx1.tissue results
}
}
\keyword{datasets}
