% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{design.rnaseq}
\alias{design.rnaseq}
\title{Design matrix for cell types (RNAseq)}
\format{
An object of class \code{matrix} with 64 rows and 4 columns.
}
\usage{
design.rnaseq
}
\description{
Design matrix for cell types (RNAseq)
}
\keyword{datasets}
