% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{dge.celltypes}
\alias{dge.celltypes}
\title{DGE object (from EdgeR) of RNAseq results for cell types}
\format{
An object of class \code{DGEList} with 500 rows and 64 columns.
}
\usage{
dge.celltypes
}
\description{
DGE object (from EdgeR) of RNAseq results for cell types
}
\keyword{datasets}
