% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/withDoc6.R
=======
% Please edit documentation in R/withDoc.R
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
\docType{data}
\name{dge.tissue}
\alias{dge.tissue}
\title{DGE object (from EdgeR) of RNAseq results for tissue}
\format{An object of class \code{DGEList} with 500 rows and 35 columns.}
\usage{
dge.tissue
}
\description{
DGE object (from EdgeR) of RNAseq results for tissue
}
\keyword{datasets}
