% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc (Autosaved).R, R/withDoc.R
\name{getb0.biseq}
\alias{getb0.biseq}
\title{compute b0, projection matrix, given methylation counts}
\usage{
getb0.biseq(methmat, design, sigg = NULL)

getb0.biseq(methmat, design, sigg = NULL)
}
\arguments{
\item{methmat}{a matrix of counts with rows corresponding to methylation sites and columns. Columns are: chr start   end strand coverage1 numCs1 numTs1 coverage2 numCs2 numTs2 ....}

\item{design}{design matrix, output of model.matrix}

\item{sigg}{predetermined signature CpG sites. Format sites as chromosome
name, chromosome location, strand: chrN_position(+/-). For example, chr1_906825-}
}
\value{
b0 projection matrix. Coefficients are beta.

b0 projection matrix. Coefficients are beta.
}
\description{
compute b0, projection matrix, given methylation counts

compute b0, projection matrix, given methylation counts
}
