% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/withDoc6.R
=======
% Please edit documentation in R/withDoc.R
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
\name{getb0.rnaseq}
\alias{getb0.rnaseq}
\title{compute b0, projection matrix, given counts}
\usage{
<<<<<<< HEAD
getb0.rnaseq(dge, design, ncpm.min = 1, nsamp.min = 4, sigg = NULL)
=======
getb0.rnaseq(dge, design, ncpm.min = 1, nsamp.min = 4)
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
}
\arguments{
\item{dge}{DGE object}

\item{design}{output of model.matrix}

\item{ncpm.min}{filtering threshold for cpm}

\item{nsamp.min}{minimum number of samples that must have cpm
exceeding ncpm.min for a gene to be retained}
<<<<<<< HEAD

\item{sigg}{vector of predetermined significant genes}
=======
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
}
\value{
b0 projection matrix

p0 pvalues matrix

dge DGE object

converged convergence status. 1=converged.

pfstat F-stat p-values for each gene
}
\description{
compute b0, projection matrix, given counts
}
