% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\name{getb0.rnaseq}
\alias{getb0.rnaseq}
\title{compute b0, projection matrix, given counts}
\usage{
getb0.rnaseq(dge, design, ncpm.min = 1, nsamp.min = 4, sigg = NULL)
}
\arguments{
\item{dge}{DGE object}

\item{design}{output of model.matrix}

\item{ncpm.min}{filtering threshold for cpm}

\item{nsamp.min}{minimum number of samples that must have cpm
exceeding ncpm.min for a gene to be retained}

\item{sigg}{vector of predetermined significant genes}
}
\value{
b0 projection matrix

p0 pvalues matrix

dge DGE object

converged convergence status. 1=converged.

pfstat F-stat p-values for each gene
}
\description{
compute b0, projection matrix, given counts
}
\examples{
\dontrun{
 data("data_celltypes_rnaseq") 
 dge.celltypes = getdge(cnts.celltypes, design.rnaseq, ncpm.min=1, nsamp.min=4)
 b0 = getb0.rnaseq(dge.celltypes, design.rnaseq, ncpm.min=1, nsamp.min=4, sigg=NULL)
 }
}
