% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\name{getcorr}
\alias{getcorr}
\title{compute correlation between estimated and actual composition for each sample}
\usage{
getcorr(x1, x2)
}
\arguments{
\item{x1}{composition matrix where rows are samples and columns are cell types}

\item{x2}{second composition matrix where rows and columns match x1}
}
\value{
correlation by sample
}
\description{
compute correlation between estimated and actual composition for each sample
}
