% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconvSeq2.R
\name{getcorr}
\alias{getcorr}
\title{compute correlation between estimated and actual composition}
\usage{
getcorr(x1, x2)
}
\arguments{
\item{x1}{composition matrix where rows are samples and columns are cell types}

\item{x2}{second composition matrix where rows and columns match x1}
}
\value{
correlation by sample
}
\description{
compute correlation between estimated and actual composition
}
\examples{
#RNA sequencing example

set.seed(1234)
data("data_celltypes_rnaseq") #contains DGE object dge.celltypes, count matrix cnts.celltypes, and design matrix des for cell types design.rnaseq, sample.id.rnaseq
dge.celltypes = getdge(cnts.celltypes, design.rnaseq, ncpm.min=1, nsamp.min=4)
b0 = getb0.rnaseq(dge.celltypes, design.rnaseq, ncpm.min=1, nsamp.min=4)
nb0 = 50
resultx1 = getx1.rnaseq(nb0,b0,dge.celltypes)
x2 = as.data.frame(design.rnaseq,row.names=sample.id.rnaseq)
getcorr(resultx1$x1,x2)

#Validate projection matrix on whole tissue
data("data_tissue_rnaseq") #contains DGE object dge.tissue, count matrix cnts.tissue, and CBC cbc.rnaseq
nb0=50
dge.tissue = getdge(cnts.tissue,design=NULL, ncpm.min=1, nsamp.min=4)
resultx1.tissue = getx1.rnaseq(nb0,b0, dge.tissue)
x1 = cbind(lymph=resultx1.tissue$x1$tissuefBcell+resultx1.tissue$x1$tissuefTcell, mono = resultx1.tissue$x1$tissuefMonocytes, gran = resultx1.tissue$x1$tissuefGranulocytes)
x2 = as.matrix(cbc.rnaseq/100)
getcorr(x1,x2)

#Bisulfite sequencing example

set.seed(1234)
data("data_celltypes_rrbs") #contains methylation matrix methmat, design.rrbs, celltypes.rrbs, sample.id.rrbs
b0 = getb0.biseq(methmat, design.rrbs)
nb0=250
resultx1 = getx1.biseq(nb0,b0,methmat,sample.id.rrbs,celltypes.rrbs)
x2 = as.data.frame(design.rrbs,row.names=sample.id.rrbs)
getcorr(resultx1$x1,x2)

#Validate projection matrix on whole tissue
data("data_tissue_rrbs") #contains methylation matrix methmat.tissue, sample.id.tissue, actual differential cell count cbc.rrbs
nb0=250
resultx1.tissue = getx1.biseq(nb0,b0,methmat.tissue,sample.id.tissue,celltypes.rrbs)
x1 = cbind(lymph=resultx1.tissue$x1[,1]+resultx1.tissue$x1[,2], mono = resultx1.tissue$x1[,3], gran = resultx1.tissue$x1[,4])
x2 = as.matrix(cbc.rrbs/100)
getcorr(x1,x2)
}
