% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc2.R
\name{getcorr}
\alias{getcorr}
\title{compute correlation between estimated and actual composition}
\usage{
getcorr(x1, x2)
}
\arguments{
\item{x1}{composition matrix where rows are samples and columns are cell types}

\item{x2}{second composition matrix where rows and columns match x1}
}
\value{
cor.type correlation by cell type

cor.sample correlation by sample
}
\description{
compute correlation between estimated and actual composition
}
\examples{
data("leucegene_btmg_b0_counts")
okids = sample.idVec()
cnts = leucegene_btmg_b0_counts[, okids]
tissf = factor(tissueVec(), levels=celltypeVec())
des = model.matrix(~-1+tissf)
b0 = getb0(cnts, des, ncpm.min=1, nsamp.min=4)
nb0 = 3000
dge = getdge(cnts, des, ncpm.min=1, nsamp.min=4)
resultx1 = getx1(nb0,b0,dge)
x2 = as.data.frame(des,row.names=okids)
result.corr = getcorr(resultx1$x1,x2)
}
