% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\name{getcorr}
\alias{getcorr}
\title{compute correlation between estimated and actual composition for each sample}
\usage{
getcorr(x1, x2)
}
\arguments{
\item{x1}{composition matrix where rows are samples and columns are cell types}

\item{x2}{second composition matrix where rows and columns match x1}
}
\value{
correlation by sample
}
\description{
compute correlation between estimated and actual composition for each sample
}
\examples{
#RNA sequencing example
#---------------------------------------------

set.seed(1234)
data("data_celltypes_rnaseq") 
dge.celltypes = getdge(cnts.celltypes, design.rnaseq, ncpm.min=1, nsamp.min=4)
b0 = getb0.rnaseq(dge.celltypes, design.rnaseq, ncpm.min=1, nsamp.min=4)
nb0 = 50
resultx1 = getx1.rnaseq(nb0,b0,dge.celltypes)
x2 = as.data.frame(design.rnaseq,row.names=sample.id.rnaseq)
getcorr(resultx1$x1,x2)

#Validate projection matrix on whole tissue
data("data_tissue_rnaseq") 
nb0=50
dge.tissue = getdge(cnts.tissue,design=NULL, ncpm.min=1, nsamp.min=4)
resultx1.tissue = getx1.rnaseq(nb0,b0, dge.tissue)
x1 = cbind(lymph=resultx1.tissue$x1$tissuefBcell+resultx1.tissue$x1$tissuefTcell, 
  mono = resultx1.tissue$x1$tissuefMonocytes, gran = resultx1.tissue$x1$tissuefGranulocytes)
x2 = as.matrix(cbc.rnaseq/100)
sc = getcorr(x1,x2)
getmeancorr(sc)


#Single cell RNA sequencing example
#---------------------------------------------

set.seed(1234)
data("data_scrnaseq") 
#Do quality control. cnts.sc and cnts.sc.G1 are included in data_scrnaseq.
#cnts.sc = prep_scrnaseq(cnts.scrnaseq, genenametype = "hgnc_symbol",cellcycle=NULL,
#count.threshold=0.05)
#Filter for cell cycle phase "G1"
#cnts.sc.G1 = getcellcycle(cnts.sc,"G1")
#divide data into a training set and a validation set
cnts.sc.G1.train = cnts.sc.G1[,c(which(substr(colnames(cnts.sc.G1),3,6)=="Tcon")[1:250],
which(substr(colnames(cnts.sc.G1),3,6)=="Treg")[1:150])]
cnts.sc.G1.valid = cnts.sc.G1[,-which(colnames(cnts.sc.G1) \%in\% colnames(cnts.sc.G1.train))]
tissue.sc = substr(colnames(cnts.sc.G1.train),3,6)
names(tissue.sc) = colnames(cnts.sc.G1.train)
sample.id.sc = colnames(cnts.sc.G1.train)
design.sc = model.matrix(~-1+as.factor(tissue.sc))
colnames(design.sc) = levels(as.factor(tissue.sc))
rownames(design.sc) = names(tissue.sc)
design.sc = design.sc[colnames(cnts.sc.G1.train),]
dge.sc = getdge(cnts.sc.G1.train,design.sc,ncpm.min=1, nsamp.min=4, method="bin.loess")
b0.sc = getb0.rnaseq(dge.sc, design.sc, ncpm.min=1, nsamp.min=4)
nb0=1500
resultx1.sc = getx1.rnaseq(nb0,b0.sc,dge.sc)
x2 = as.data.frame(design.sc,row.names=sample.id.sc)
sc = getcorr(resultx1.sc$x1,x2)
getmeancorr(sc)

#Apply projection matrix to validation set
tissue_s.sc = substr(colnames(cnts.sc.G1.valid),3,6)
names(tissue_s.sc) = colnames(cnts.sc.G1.valid)
sample.id_s.sc = colnames(cnts.sc.G1.valid)
design_s.sc = model.matrix(~-1+as.factor(tissue_s.sc))
colnames(design_s.sc) = levels(as.factor(tissue_s.sc))
rownames(design_s.sc) = names(tissue_s.sc)
design_s.sc = design_s.sc[colnames(cnts.sc.G1.valid),]
dge_s.sc = getdge(cnts.sc.G1.valid, design_s.sc, ncpm.min=1, nsamp.min=4, method="bin.loess")
resultx1_s.sc = getx1.rnaseq(nb0,b0.sc, dge_s.sc)
x2 = as.data.frame(design_s.sc,row.names=sample.id_s.sc)
sc = getcorr(resultx1_s.sc$x1,x2)
getmeancorr(sc)


#Bisulfite sequencing example
#---------------------------------------------

set.seed(1234)
data("data_celltypes_rrbs") 
b0 = getb0.biseq(methmat, design.rrbs, sigg=NULL)
nb0=250
resultx1 = getx1.biseq(nb0,b0,methmat,sample.id.rrbs,celltypes.rrbs)
x2 = as.data.frame(design.rrbs,row.names=sample.id.rrbs)
getcorr(resultx1$x1,x2)

#Validate projection matrix on whole tissue
data("data_tissue_rrbs") 
nb0=250
resultx1.tissue = getx1.biseq(nb0,b0,methmat.tissue,sample.id.tissue,celltypes.rrbs)
x1 = cbind(lymph=resultx1.tissue$x1[,1]+resultx1.tissue$x1[,2], 
  mono = resultx1.tissue$x1[,3], gran = resultx1.tissue$x1[,4])
x2 = as.matrix(cbc.rrbs/100)
sc = getcorr(x1,x2)
getmeancorr(sc)
}
