% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconvSeq2.R, R/withDoc.R
\name{getdge}
\alias{getdge}
\alias{getdge}
\title{obtain EdgeR DGE object from count matrix}
\usage{
getdge(countmat, design, ncpm.min, nsamp.min)

getdge(countmat, design, ncpm.min, nsamp.min)
}
\arguments{
\item{countmat}{a matrix of counts with rows corresponding to genes and columns to samples}

\item{design}{output of model.matrix}

\item{ncpm.min}{filtering threshold for cpm}

\item{nsamp.min}{minimum number of samples that must have cpm exceeding ncpm.min for a gene to be retained}

\item{countmat}{a matrix of counts with rows corresponding to genes and columns to samples}

\item{design}{output of model.matrix}

\item{ncpm.min}{filtering threshold for cpm}

\item{nsamp.min}{minimum number of samples that must have cpm exceeding ncpm.min for a gene to be retained}
}
\value{
dge DGE object

dge DGE object
}
\description{
obtain EdgeR DGE object from count matrix

obtain EdgeR DGE object from count matrix
}
