% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\name{getdge}
\alias{getdge}
\title{Obtain EdgeR DGE object from count matrix}
\usage{
getdge(countmat, design, ncpm.min = 1, nsamp.min = 4, method = "auto")
}
\arguments{
\item{countmat}{a matrix of counts with rows corresponding to
genes and columns to samples}

\item{design}{output of model.matrix}

\item{ncpm.min}{filtering threshold for cpm}

\item{nsamp.min}{minimum number of samples that must have cpm}

\item{method}{method used to esimate trended dispersion in EdgeR: "auto" (default), "bin.spline", "bin.loess", "power", "spline"
exceeding ncpm.min for a gene to be retained}
}
\value{
dge DGE object
}
\description{
Obtain EdgeR DGE object from count matrix
}
\examples{
data("data_celltypes_rnaseq") 
dge.celltypes = getdge(cnts.celltypes, design.rnaseq, ncpm.min=1, nsamp.min=4)
}
