% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc6.R
\name{getmeancorr}
\alias{getmeancorr}
\title{get mean correlation by first doing Fisher transform to z, averaging z, then reverse transform back
do quality control of scRNAseq based on library size, feature counts, chrM, cell cycle phase, and count threshold}
\usage{
getmeancorr(rho)
}
\arguments{
\item{rho}{correlations}
}
\value{
mean correlation and 95% CI
}
\description{
get mean correlation by first doing Fisher transform to z, averaging z, then reverse transform back
do quality control of scRNAseq based on library size, feature counts, chrM, cell cycle phase, and count threshold
}
