% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\name{getmethmat}
\alias{getmethmat}
\title{Makes methylation count matrix}
\usage{
getmethmat(filnames, sample.id, filtype = "bsmap")
}
\arguments{
\item{filnames}{input filenames}

\item{sample.id}{sample IDs}

\item{filtype}{"bsmap", "bismark". Default is "bsmap"}
}
\value{
methylation count matrix where rows are CpG sites and columns are: chromosome, start, end, strand, number of Ts+Cs for sample 1, number of Cs for sample 1, number of Ts for sample 1, ....
}
\description{
Makes methylation count matrix using output from either BSMAP or Bismark.
}
\details{
Removes sites with 0 counts in all samples.

Input file from BSMAP has columns: chromosome, position, strand, context, ratio, eff_CT_count, C_count, CT_count, rev_G_count, rev_GA_count, CI_lower, CI_upper.

Input file from Bismark coverage for CG only has columns "chrBase","chr","base","strand","coverage","freqC","freqT"

Sample input files from BSMAP: sample1_methratio.txt, sample1_mehtratio.txt

For BSMAP input files, extracts CpGs and output in filname.CpG in the current working directory.
Adds "Chr" to chromosome name and output in filname.CpG_chr.
}
\examples{
 file1 = system.file("extdata","sample1_methratio.txt", package="deconvSeq")
 file2 = system.file("extdata","sample2_methratio.txt", package="deconvSeq")
methmat = getmethmat(filnames=c(file1,file2), sample.id=c("sample1","sample2"), filtype="bsmap")
}
