% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconvSeq2.R
\name{getx1.biseq}
\alias{getx1.biseq}
\title{compute mixture data given projection matrix (bisulfite sequencing)}
\usage{
getx1.biseq(NB0, b0, methmat, sample.id, celltypes, MAXITER = 10000,
  x0 = NULL)
}
\arguments{
\item{NB0}{number of genes to be retained, ordered by f-stat p-value}

\item{b0}{output of \code{\link{getb0.biseq}}}

\item{methmat}{a matrix of counts with rows corresponding to methylation sites and columns. Columns are: chr start   end strand coverage1 numCs1 numTs1 coverage2 numCs2 numTs2 ....}

\item{sample.id}{vector of sample IDs}

\item{celltypes}{vector of cell types}

\item{MAXITER}{integer number of iterations allowed}

\item{x0}{initial cell type composition for fitting}
}
\value{
x1 cell mixture of sample

converged convergence
}
\description{
compute mixture data given projection matrix (bisulfite sequencing)
}
