% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/withDoc6.R
=======
% Please edit documentation in R/withDoc.R
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
\name{getx1.rnaseq}
\alias{getx1.rnaseq}
\title{compute mixture data given projection matrix (RNAseq)}
\usage{
<<<<<<< HEAD
getx1.rnaseq(NB0 = "top_bonferroni", resultb0, dge_s, MAXITER = 1000,
  x0 = NULL)
}
\arguments{
\item{NB0}{number of genes to be retained, ordered by f-stat p-value. Other options: "all" uses all genes, "top_bonferroni" uses genes with adjusted p-values <0.05 after bonferroni correction, "top_fdr" uses genes with adjusted p-values <0.05 after FDR correction. Default is "top-bonferroni"}
=======
getx1.rnaseq(NB0, resultb0, dge_s, MAXITER = 1000, x0 = NULL)
}
\arguments{
\item{NB0}{number of genes to be retained, ordered by f-stat p-value}
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555

\item{resultb0}{output of \code{\link{getb0.rnaseq}}}

\item{dge_s}{output of \code{\link{getdge}}}

\item{MAXITER}{integer number of iterations allowed}

\item{x0}{initial cell type composition for fitting}
}
\value{
x1 cell mixture of sample

converged convergence. 1=converged
}
\description{
compute mixture data given projection matrix (RNAseq)
}
