% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/withDoc6.R
=======
% Please edit documentation in R/withDoc.R
>>>>>>> da6dabc5622e35b9fed80bedbbcd472d3a979555
\docType{data}
\name{methmat.tissue}
\alias{methmat.tissue}
\title{methylation matrix for tissue (bisulfite sequencing)}
\format{An object of class \code{data.frame} with 500 rows and 109 columns.}
\usage{
methmat.tissue
}
\description{
methylation matrix for tissue (bisulfite sequencing)
}
\keyword{datasets}
