% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{methmat.tissue}
\alias{methmat.tissue}
\title{methylation matrix for tissue (bisulfite sequencing)}
\format{
An object of class \code{data.frame} with 500 rows and 109 columns.
}
\usage{
methmat.tissue
}
\description{
methylation matrix for tissue (bisulfite sequencing)
}
\keyword{datasets}
