% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc.R
\docType{data}
\name{sample.id.rnaseq}
\alias{sample.id.rnaseq}
\title{Sample IDs for cell types (RNAseq)}
\format{
An object of class \code{character} of length 64.
}
\usage{
sample.id.rnaseq
}
\description{
Sample IDs for cell types (RNAseq)
}
\keyword{datasets}
