% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc (Autosaved).R, R/withDoc.R
\docType{data}
\name{sample.id.rrbs}
\alias{sample.id.rrbs}
\title{Sample IDs for cell types (bisulfite sequencing)}
\format{
An object of class \code{character} of length 24.

An object of class \code{character} of length 24.
}
\usage{
sample.id.rrbs

sample.id.rrbs
}
\description{
Sample IDs for cell types (bisulfite sequencing)

Sample IDs for cell types (bisulfite sequencing)
}
\keyword{datasets}
