% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withDoc (Autosaved).R, R/withDoc.R
\docType{data}
\name{sample.id.tissue}
\alias{sample.id.tissue}
\title{Sample IDs for tissue (bisulfite sequencing)}
\format{
An object of class \code{character} of length 35.

An object of class \code{character} of length 35.
}
\usage{
sample.id.tissue

sample.id.tissue
}
\description{
Sample IDs for tissue (bisulfite sequencing)

Sample IDs for tissue (bisulfite sequencing)
}
\keyword{datasets}
